enable permalinks
This extracts a method to add elements to the plot, which in turn enables us to manage permalinks by calling that function with parameters from the URL hash. Also includes a permalink html element which will set the document hash to a value that can then be shared. Fixes issue #6